+++ src/gwt/src/org/rstudio/studio/client/server/remote/RemoteServer.java
@@ -64,6 +64,7 @@ import org.rstudio.studio.client.common.debugging.model.FunctionSteps;
 import org.rstudio.studio.client.common.debugging.model.TopLevelLineData;
 import org.rstudio.studio.client.common.dependencies.model.Dependency;
 import org.rstudio.studio.client.common.mirrors.model.CRANMirror;
+import org.rstudio.studio.client.common.plumber.model.PlumberCapabilities;
 import org.rstudio.studio.client.common.presentation.model.SlideNavigation;
 import org.rstudio.studio.client.common.r.roxygen.RoxygenHelper.SetClassCall;
 import org.rstudio.studio.client.common.r.roxygen.RoxygenHelper.SetGenericCall;
@@ -93,6 +94,8 @@ import org.rstudio.studio.client.packrat.model.PackratContext;
 import org.rstudio.studio.client.packrat.model.PackratPackageAction;
 import org.rstudio.studio.client.packrat.model.PackratPrerequisites;
 import org.rstudio.studio.client.packrat.model.PackratStatus;
+import org.rstudio.studio.client.plumber.model.PlumberRunCmd;
+import org.rstudio.studio.client.plumber.model.PlumberViewerType;
 import org.rstudio.studio.client.projects.model.NewPackageOptions;
 import org.rstudio.studio.client.projects.model.NewProjectContext;
 import org.rstudio.studio.client.projects.model.NewShinyAppOptions;
@@ -1676,6 +1679,16 @@ public class RemoteServer implements Server
       sendRequest(RPC_SCOPE, CREATE_SHINY_APP, params, requestCallback);
    }
    
+   public void createPlumberAPI(String apiName,
+                                String apiDir,
+                                ServerRequestCallback<JsArrayString> requestCallback)
+   {
+      JSONArray params = new JSONArray();
+      params.set(0, new JSONString(apiName));
+      params.set(1, new JSONString(apiDir));
+      sendRequest(RPC_SCOPE, CREATE_PLUMBER_API, params, requestCallback);
+   }
+   
    public void getEditorContextCompleted(GetEditorContextEvent.SelectionData data,
                                          ServerRequestCallback<Void> requestCallback)
    {
@@ -2212,7 +2225,7 @@ public class RemoteServer implements Server
    {
       sendRequest(RPC_SCOPE, "get_shiny_capabilities", requestCallback);
    }
-
+   
    public void getRecentHistory(
          long maxItems,
          ServerRequestCallback<RpcObjectList<HistoryEntry>> requestCallback)
@@ -4299,6 +4312,26 @@ public class RemoteServer implements Server
             params,
             requestCallback);
    }
+   
+   @Override
+   public void getPlumberViewerType(ServerRequestCallback<PlumberViewerType> requestCallback)
+   {
+      sendRequest(RPC_SCOPE,
+            GET_PLUMBER_VIEWER_TYPE,
+            requestCallback);
+   }
+
+   @Override
+   public void getPlumberRunCmd(String plumberFile, 
+                                ServerRequestCallback<PlumberRunCmd> requestCallback)
+   {
+      JSONArray params = new JSONArray();
+      params.set(0, new JSONString(plumberFile));
+      sendRequest(RPC_SCOPE,
+            GET_PLUMBER_RUN_CMD,
+            params,
+            requestCallback);
+   }
 
    @Override
    public void getRSConnectAccountList(
@@ -5293,7 +5326,7 @@ public class RemoteServer implements Server
       JSONArray params = new JSONArray();
       sendRequest(RPC_SCOPE, STOP_SHINY_APP, params, true, callback);
    }
-
+  
    @Override
    public void connectionAddPackage(String packageName,
                                     ServerRequestCallback<Void> callback)
@@ -5769,6 +5802,10 @@ public class RemoteServer implements Server
    private static final String GET_SHINY_VIEWER_TYPE = "get_shiny_viewer_type";
    private static final String GET_SHINY_RUN_CMD = "get_shiny_run_cmd";
    private static final String SET_SHINY_VIEWER_TYPE = "set_shiny_viewer_type";
+   
+   private static final String CREATE_PLUMBER_API = "create_plumber_api";
+   private static final String GET_PLUMBER_VIEWER_TYPE = "get_plumber_viewer_type";
+   private static final String GET_PLUMBER_RUN_CMD = "get_plumber_run_cmd";
    
    private static final String GET_RSCONNECT_ACCOUNT_LIST = "get_rsconnect_account_list";
    private static final String REMOVE_RSCONNECT_ACCOUNT = "remove_rsconnect_account";