+++ src/gwt/src/org/rstudio/studio/client/workbench/prefs/views/PublishingPreferencesPane.java
@@ -1,7 +1,7 @@
 /*
  * PublishingPreferencesPane.java
  *
- * Copyright (C) 2009-14 by RStudio, Inc.
+ * Copyright (C) 2009-18 by RStudio, Inc.
  *
  * Unless you have received this program directly from RStudio pursuant
  * to the terms of a commercial license agreement with RStudio, then
@@ -193,7 +193,7 @@ public class PublishingPreferencesPane extends PreferencesPane
       lessSpaced(rsconnectPanel);
       
       add(headerLabel("Settings"));
-      CheckBox chkEnablePublishing = checkboxPref("Enable publishing documents and apps", 
+      CheckBox chkEnablePublishing = checkboxPref("Enable publishing documents, apps, and APIs", 
             uiPrefs_.showPublishUi());
       chkEnablePublishing.addValueChangeHandler(new ValueChangeHandler<Boolean>(){
          @Override